+++ src/main/java/ru/yandex/practicum/filmorate/storage/GenreStorage.java
@@ -12,5 +12,5 @@ public interface GenreStorage {
 
     Genre update(Genre newGenre);
 
-    Optional<Genre> findGenreById(Long idOfGenre);
+    Optional<Genre> findGenreById(int idOfGenre);
 }
